trial/requirements.txt

# Flask Backend Requirements
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Video Processing
opencv-python==4.8.1.78
moviepy==1.0.3
numpy==1.24.3

# Audio Processing & Transcription
SpeechRecognition==3.10.0
pyaudio==0.2.11
openai-whisper==20231117

# HTTP Requests
requests==2.31.0

# Utility
Pillow==10.0.1
python-dotenv==1.0.0

# Optional: For better performance
# torch==2.0.1
# torchvision==0.15.2
# torchaudio==2.0.2
